+++ .gitignore
@@ -1,4 +1,6 @@
+.idea
 pkg/
 src/
 tmp/
-alp
+cli/alp/*
+!cli/alp/main.go
